refactoring

+++ fintrack/src/main/java/kz/finance/fintrack/dto/CategoryExpenseDetailsDto.java
@@ -1,12 +1,14 @@
 package kz.finance.fintrack.dto;
 
+import kz.finance.fintrack.model.ExpenseCategory;
+
 import java.math.BigDecimal;
 import java.util.List;
 
 public record CategoryExpenseDetailsDto(
         ExpenseCategory category,
         BigDecimal amount,
-        int percentage,
+        BigDecimal percentage,
         List<ExpenseDto> expenses
 ) {
 }
